requirements.txt
colorama
dpkt
pypcap==1.3.0
rich==13.7.0
scikit_learn==1.4.0
